Add IMDb search form and integration
Introduced a new IMDb search form with advanced filters for title, plot, keywords, genres, release dates, ratings, and more. Added IMDb as an option in the search selector and implemented the searchIMDb() function to construct and open IMDb search URLs based on user input. Also updated SafeSearch options for Google Web and Images to set 'Off' as the default.